# v3/api/openapi.yaml
openapi: 3.1.0
info:
  title: Sakura v3 API
  version: 3.0.0-alpha.1
  description: Unified API used equally by the Web portal, Telegram Bot and API clients.
servers:
  - url: /
paths:
...
  /api/v3/me:
    get:
      operationId: getCurrentAccount
      tags: [Accounts]
      security: [{cookieAuth: []}]
      responses:
        "200": {description: Unified account and identities}
...
  /api/v3/admin/settings:
    get:
      operationId: listDynamicSettings
      tags: [Settings]
      security: [{cookieAuth: []}]
      responses:
        "200": {description: Current versioned settings}
  /api/v3/admin/settings/{key}:
    parameters:
      - $ref: "#/components/parameters/SettingKey"
    patch:
      operationId: updateDynamicSetting
      tags: [Settings]
      security: [{cookieAuth: [], csrfHeader: []}]
      responses:
        "200": {description: New setting revision}
        "409": {$ref: "#/components/responses/Conflict"}
  /api/v3/admin/settings/{key}/history:
    parameters:
      - $ref: "#/components/parameters/SettingKey"
    get:
      operationId: getDynamicSettingHistory
      tags: [Settings]
      security: [{cookieAuth: []}]
      responses:
        "200": {description: Immutable setting revisions}
  /api/v3/admin/settings/{key}/rollback:
    parameters:
      - $ref: "#/components/parameters/SettingKey"
    post:
      operationId: rollbackDynamicSetting
      tags: [Settings]
      security: [{cookieAuth: [], csrfHeader: []}]
      responses:
        "200": {description: Rollback recorded as a new revision}
        "409": {$ref: "#/components/responses/Conflict"}
...
  /api/v3/me/notifications:
    get:
      operationId: listMyNotifications
      tags: [Notifications]
      security: [{cookieAuth: []}]
      responses: {"200": {description: Notifications addressed to current account}}
  /api/v3/me/notifications/{id}/read:
    parameters: [{$ref: "#/components/parameters/ID"}]
    post:
      operationId: markNotificationRead
      tags: [Notifications]
      security: [{cookieAuth: [], csrfHeader: []}]
      responses: {"204": {description: Notification marked read}}
...
  /api/v3/reviews:
    get:
      operationId: listPublicReviews
      tags: [Reviews]
      security: [{cookieAuth: []}]
      responses: {"200": {description: Approved media reviews}}
  /api/v3/me/reviews:
    get:
      operationId: listMyReviews
      tags: [Reviews]
      security: [{cookieAuth: []}]
      responses: {"200": {description: Current account reviews including moderation state}}
    post:
      operationId: submitReview
      tags: [Reviews]
      security: [{cookieAuth: [], csrfHeader: []}]
      responses: {"201": {description: Review submitted for moderation}}
  /api/v3/admin/reviews:
    get:
      operationId: listReviewsForModeration
      tags: [Administration, Reviews]
      security: [{cookieAuth: []}]
      responses: {"200": {description: Review moderation queue}}
...
  /api/v3/me/notification-preferences:
    get:
      operationId: listNotificationPreferences
      tags: [Notifications]
      security: [{cookieAuth: []}]
      responses: {"200": {description: Effective per-event and per-channel preferences}}
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sakura_v3_session
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
...
    ID:
      name: id
      in: path
      required: true
      schema: {type: string, format: uuid}
    SettingKey:
      name: key
      in: path
      required: true
      schema: {type: string}
  responses:
...
    Conflict: {description: Revision or uniqueness conflict}
...